Create python code
There can be a variety of data present on any website. But all of those data may not be required for use. So web scraping gives us the option to select required data for any purpose. In our project, we have selected indeed.com where various jobs are listed. Here we arrange a listing of a job in form of a table. This table helps us analyze, and compare different jobs. Our specific search is made easier this way.